Factor around tests to match what they are testing re:@benwah
